:art: refactor: SearchBar 컴포넌트 분리
- SearchBar 컴포넌트를 src/app/components/SearchBar.tsx 파일로 이동합니다.
